Updated language
I switched "leader" to "alpha bat" since it seems to make the most sense when talking about animals (even if they are robots.) It might make sense to switch from "L" to "A" when marking the alpha bats location. Making the bats robot bats seems to solve the issue of why the leader would have the computer chip.